requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.60.0
matplotlib>=3.5.0
